[test] win32: terminate node http server by killing all subprocesses
The SIGTERM handler in node.py doesn't work on windows

+++ test/sanity/issue4877-inject-webview/test.py
@@ -45,4 +45,8 @@ try:
     assert(len(elems) == 1)
 finally:
     driver.quit()
-    server.terminate()
+    import platform
+    if platform.system() == 'Windows':
+        subprocess.call(['taskkill', '/F', '/T', '/PID', str(server.pid)])
+    else:
+        server.terminate()
